Fix glitchy filter buttons on workbench (bsc#956803)
Somehow, not having the font-size in the .make-switch class match the
default body font size seems to cause a click event to double fire, or
missfire somehow (it lands on a div instead of a label), which results
in it being possible to turn a filter button off, but never back on
again.

Additionally, setting the circle fill style in the SUSE branding somehow
results in the OSDs never disappearing from the workbench when filtered
off.  AFAICT this fill style doesn't do anything useful anyway, so it
seems safe to remove it.

Signed-off-by: Tim Serong <tserong@suse.com>