Remove OSD hostname check on summary graphs
When displaying summary graphs for a host, there was a check that
the host's FQDN was in the list of FQDNs of OSD hosts.  This means
summary graphs woudln't display if a MON was on a separate host
all by itself.

Signed-off-by: Tim Serong <tserong@suse.com>
(cherry picked from commit d2e51c893b751c9f9168374a967d3fff30f4a366)